I legit do not know why it was a ushort.

+++ VoiceCraft.Core/Packets/MCComm/GetChannels.cs
@@ -5,6 +5,6 @@ namespace VoiceCraft.Core.Packets.MCComm
     public class GetChannels : MCCommPacket
     {
         public override byte PacketId => (byte)MCCommPacketTypes.GetChannels;
-        public Dictionary<ushort, Channel> Channels { get; set; } = new Dictionary<ushort, Channel>();
+        public Dictionary<byte, Channel> Channels { get; set; } = new Dictionary<byte, Channel>();
     }
 }
